+++ src/ui/AMultiselectOrdered/AMultiselectOrdered.js
@@ -124,6 +124,11 @@ export default {
       required: false,
       default: () => [],
     },
+    mergeData: {
+      type: Boolean,
+      required: false,
+      default: false,
+    },
     dependencies: {
       type: [Array, Object],
       required: false,