+++ src/ui/ARadio/ARadio.js
@@ -93,6 +93,11 @@ export default {
       required: false,
       default: () => [],
     },
+    mergeData: {
+      type: Boolean,
+      required: false,
+      default: false,
+    },
     dependencies: {
       type: [Array, Object],
       required: false,